refactor: move get downtime counter function
